refactor: extract stridedview-rs crate from strided-rs (#11)
Separate device-agnostic view types and metadata operations into a
standalone stridedview-rs crate, following Julia's StridedViews.jl /
Strided.jl separation. This enables downstream crates to depend on
view types without pulling in the full kernel engine.

- Convert to Cargo workspace with stridedview/ and strided/ members
- Move StridedView, StridedViewMut, StridedArray, ElementOp types
  and auxiliary helpers to stridedview-rs
- Move StridedError and Result<T> to stridedview-rs
- strided-rs re-exports all stridedview types for backward compatibility
- No public API changes for existing strided-rs users
