+++ strided/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "strided-rs"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+authors = ["Satoshi Terasaki", "Hiroshi Shinaoka"]
+repository = "https://github.com/tensor4all/strided-rs"
+description = "Cache-optimized kernels for strided multidimensional array operations in Rust (ported from Julia Strided.jl/StridedViews.jl)."
+publish = false
+
+[dependencies]
+stridedview = { path = "../stridedview", package = "stridedview-rs" }
+num-traits = "0.2"
+num-complex = "0.4"
+rayon = { version = "1.10", optional = true }
+
+[features]
+default = []
+parallel = ["rayon"]
+
+[dev-dependencies]
+criterion = "0.5"
+approx = "0.5"
+rand = "0.8"
+rand_distr = "0.4"
+
+[lib]
+bench = false
+
+[[bench]]
+name = "rust_compare"
+harness = false
+
+[[bench]]
+name = "scaling_compare"
+harness = false
+
+[[bench]]
+name = "threaded_compare"
+harness = false
+required-features = ["parallel"]
