+++ strided/benches/julia_compare.jl
