+++ strided/benches/run_all.sh
