+++ strided/benches/run_scaling.sh
