+++ strided/benches/run_threaded.sh
