+++ strided/benches/rust_compare.rs
