+++ strided/src/kernel.rs
