+++ strided/src/map_view.rs
@@ -2,13 +2,13 @@
 //!
 //! These are the canonical view-based map functions, equivalent to Julia's `Base.map!`.
 
-use crate::element_op::{ElementOp, ElementOpApply};
 use crate::kernel::{
     build_plan_fused, ensure_same_shape, for_each_inner_block_preordered, is_contiguous, total_len,
     use_sequential_fast_path,
 };
 use crate::strided_view::{StridedView, StridedViewMut};
 use crate::Result;
+use stridedview::{ElementOp, ElementOpApply};
 
 #[cfg(feature = "parallel")]
 use crate::threading::{
