+++ strided/src/reduce_view.rs
@@ -1,12 +1,12 @@
 //! Reduce operations on dynamic-rank strided views.
 
-use crate::element_op::{ElementOp, ElementOpApply};
 use crate::kernel::{
     build_plan_fused, for_each_inner_block_preordered, is_contiguous, total_len,
     use_sequential_fast_path,
 };
 use crate::strided_view::{col_major_strides, StridedArray, StridedView};
 use crate::{Result, StridedError};
+use stridedview::{ElementOp, ElementOpApply};
 
 #[cfg(feature = "parallel")]
 use crate::threading::{
