+++ stridedview/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "stridedview-rs"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+authors = ["Satoshi Terasaki", "Hiroshi Shinaoka"]
+repository = "https://github.com/tensor4all/strided-rs"
+description = "Device-agnostic strided view types and metadata operations (ported from Julia StridedViews.jl)."
+publish = false
+
+[lib]
+name = "stridedview"
+
+[dependencies]
+num-traits = "0.2"
+num-complex = "0.4"
+thiserror = "1.0"
