+++ stridedview/src/auxiliary.rs
